add all endpoints working! starting on jwt
