+++ routes/rankings.js
@@ -2,8 +2,10 @@ const express = require('express');
 const api = express.Router();
 
 //controllers
-const getRanking = require('../controllers/rankings.js');
+const {getRanking, getWinner, getLoser} = require('../controllers/rankings.js');
 
 api.get('/rankings', getRanking);
+api.get('/rankings/loser', getLoser);
+api.get('/rankings/winner', getWinner);
 
 module.exports = api;
