+++ utils/auth.js
@@ -1,0 +1,1 @@
+const jwt = require('jsonwebtoken');
